+++ src/frontend/app/components/app/app.component.spec.ts
@@ -16,6 +16,11 @@ import {PlatformService} from '@dps/mycms-frontend-commons/dist/angular-commons/
 import {PageUtils} from '@dps/mycms-frontend-commons/dist/angular-commons/services/page.utils';
 import {LayoutService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/layout.service';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
+import {PrintService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/print.service';
+import {SimplePrintService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/simple-print.service';
+import {PdfGenerator, PdfPrintService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/pdf-print.service';
+import {PrintDialogPdfGenerator} from '@dps/mycms-frontend-commons/dist/angular-commons/services/print-dialog-pdf.generator';
+import {SimplePdfPrintService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/simple-pdf-print.service';
 
 describe('AppComponent', () => {
     beforeEach(() => {
@@ -37,7 +42,10 @@ describe('AppComponent', () => {
                 { provide: ToastrService, useValue: new ToastrServiceStub() },
                 PlatformService,
                 PageUtils,
-                LayoutService
+                LayoutService,
+                {provide: PrintService, useClass: SimplePrintService},
+                {provide: PdfGenerator, useClass: PrintDialogPdfGenerator},
+                {provide: PdfPrintService, useClass: SimplePdfPrintService}
             ],
             schemas: [NO_ERRORS_SCHEMA]
         });